% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Group stansim objects into a collection}
\usage{
collect(collection_name, object, ...)
}
\arguments{
\item{collection_name}{A name attached to the \code{stansim_collection}
object to help identify it. It is strongly recomended that an informative
name is assigned.}

\item{object}{An object of classs \code{stansim_simulation} or
\code{stansim_collection}}

\item{...}{Any further \code{stansim_simulation} or
\code{stansim_collection} objects to be grouped into a single
\code{stansim_collection} object.}
}
\value{
An S3 object of class \code{stansim_simulation} recording relevant
simulation data.
}
\description{
\code{collect} groups together an arbitrary number of objects
with class \code{stansim_simulation} or \code{stansim_collection} into a
single \code{stansim_collection} object. Allows for multiple simulations
to be stored, saved, analysed and managed in a single object.
}
