% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal functions.R
\name{coresCalc}
\alias{coresCalc}
\title{Calculate number of multi-chain models that can be fit simultaneously}
\usage{
coresCalc(stanChains = 1L, useCores = 1L)
}
\arguments{
\item{stanChains}{The number of chains to run in each stan model}

\item{useCores}{The total number of cores to use in the simulation}
}
\value{
The number of models that can be estimated in parallel
}
\description{
\code{coresCalc} Takes the number of chains to run per stan model, and the
total number of cores to be used in the simulation. From these it
calculates the number of instances of \code{safeFit} that can be ran
concurrently.
}
\examples{
coresCalc(4L, 8L)

}
