% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansimObjects.R
\name{extract_data.stansim}
\alias{extract_data.stansim}
\title{Extract data from a stansim object}
\usage{
\method{extract_data}{stansim}(object, dataset = "all", parameter = "all",
  estimate = "all", value = NULL)
}
\arguments{
\item{x}{An object of S3 class stansim.}
}
\description{
get data out of a single stansim object
}
