% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_refitted-methods.R
\name{extract_refitted}
\alias{extract_refitted}
\title{Extract identifiers of refitted items from rstansim objects}
\usage{
extract_refitted(object, ...)
}
\arguments{
\item{object}{An S3 object of class stansim_simulation or
stansim_collection.}

\item{...}{Arguments for filtering returned refitted records,
see specific methods for further detail.}
}
\value{
The specified refitted identifiers,
see specific methods for further detail.
}
\description{
Generic function for extracting refitted identifiers from
rstansim objects. Default arguments will return full data as a dataframe,
otherwise rows will be filtered based on provided arguments
(stansim_collections only).
}
