% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal functions.R
\name{Log_likCheck}
\alias{Log_likCheck}
\title{Checks that the user has specified a log_lik object in their stan code}
\usage{
Log_likCheck(stanModel)
}
\arguments{
\item{stanModel}{TThe file location of a valid stan model}
}
\value{
Either passes nothing if succesful or results in an error
}
\description{
\code{log_likCheck} uses a regular expression to test, in a very broad way,
whether the users stan model has a lok_lik object specified in the generated
quantities block of the code. This is required if LOO statistics are to
be calculated. If it's existence can not be confirmed users are prompted
whether they wish to continue with the simulation or stop it.
}
