% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansimObjects.R
\name{print.stansim}
\alias{print.stansim}
\title{Print a summary for a stansim simulation object}
\usage{
\method{print}{stansim}(x)
}
\arguments{
\item{x}{An object of S3 class stansim.}
}
\description{
Print basic  information regarding a stansim simulation object,
including simulation title, time/date ran, number of models fitted, and parameters
and estimates recorded.
}
