% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.stansim_collection}
\alias{print.stansim_collection}
\title{Print a summary for a stansim_collection object}
\usage{
\method{print}{stansim_collection}(x, ...)
}
\arguments{
\item{x}{An object of S3 class stansim_collection.}

\item{...}{other arguments not used by this method}
}
\description{
Print basic information regarding a stansim_collection object,
  including collection title, the simulations within the collection, and any
  datafiles that have been refitted.
}
\examples{
\dontrun{
# print stansim_collection summary info
print(collection)
}

}
\seealso{
S3 class \code{\link{stansim_collection}}.
}
