% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.stansim_data}
\alias{print.stansim_data}
\title{Print a summary for a stansim_data object}
\usage{
\method{print}{stansim_data}(x, ...)
}
\arguments{
\item{x}{An object of S3 class stansim_data.}

\item{...}{other arguments not used by this method}
}
\description{
Print basic information regarding a stansim_data object,
  including data title, number of datasets, variables recorded, and the name
  of the model from which the data was simulated from.
}
\examples{
\dontrun{
# print stansim_data summary info
print(simulated_data)
}

}
\seealso{
S3 class \code{\link{stansim_data}}.
}
