% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitMethods.R
\name{refit.stansim_single}
\alias{refit.stansim_single}
\title{Refit specified datafiles in a stansim_single object}
\usage{
\method{refit}{stansim_single}(object, datafiles = "all", all_warn = TRUE)
}
\arguments{
\item{object}{A list of function arguments to be used by
the internal \code{stan} function when fitting the models.
If not specified then the \code{stan} function defaults are used.}

\item{sim_data}{A list of strings pointing to the location of
.rds files containing the simulation data. See the vignette on
producing simulation data for details on the formatting of these datasets.}

\item{calc_loo}{If \code{TRUE} then model fit statsics will be
calculated using the \code{loo} package. If \code{TRUE} there must be
a valid log_lik quantity specified in the generated quantities
section of the provided stan model.}
}
\value{
An S3 object of class \code{stansim_single} recording relevant
simulation data.
}
\description{
\code{refit}
}
