% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-methods.R
\name{rename}
\alias{rename}
\title{Rename a stansim object}
\usage{
rename(object, new_name)
}
\arguments{
\item{object}{An S3 object of class stansim_simulation or stansim_collection.}

\item{new_name}{New object name.}
}
\value{
A stansim S3 object.
}
\description{
Generic function for renaming rstansim objects.
}
