% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-methods.R
\name{rename.stansim_collection}
\alias{rename.stansim_collection}
\title{Rename a stansim_collection object}
\usage{
\method{rename}{stansim_collection}(object, new_name)
}
\arguments{
\item{object}{An object of S3 class stansim_collection.}

\item{new_name}{New object name. Must be of type character.}
}
\value{
An S3 object of class \code{stansim_collection}.
}
\description{
Change the collection_name value of a \
  code{"stansim_collection"} object.
}
\examples{
\dontrun{
# rename a stansim_collection to "new collection name"
rename(collection, "new collection name")
}

}
