% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-methods.R
\name{rename.stansim_simulation}
\alias{rename.stansim_simulation}
\title{Rename a stansim_simulation object}
\usage{
\method{rename}{stansim_simulation}(object, new_name)
}
\arguments{
\item{object}{An object of S3 class stansim_simulation.}

\item{new_name}{New object name. Must be of type character.}
}
\description{
Change the sim_name value of a \code{"stansim_simulation"}
object.
}
