% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstansim.R
\docType{package}
\name{rstansim}
\alias{rstansim}
\alias{rstansim-package}
\title{rstansim: running simulation studies with Stan}
\description{
rstansim provides a set of helper functions and utilities to make it easier
to setup, run, analyse, and manage simulation studies in R that use
Stan for estimating Bayesian models.
}
\details{
The package addresses three aspects of running a simulation study:

\itemize{
\item Data simulation [currently not covered].
\item Model fitting and capture of relevant data.
\item Management of simulation results.
}

To learn more about rstansim, start with the vignettes:
`browseVignettes(package = "rstansim")`
}
