% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{safeFit}
\alias{safeFit}
\title{Fit a stan model to data testing for convergence}
\usage{
safeFit(stanModel, stanData, stanIter = 5000, stanChains = 4,
  maxRhat = 1.05, maxFailure = 5)
}
\arguments{
\item{stanModel}{The file location of a valid stan model}

\item{stanData}{All necessary stan data stored in a named list}

\item{stanIter}{The number of iterations to run when fitting the model}

\item{stanChains}{The number of chains to run when fitting the model}

\item{maxRhat}{The maximum R-hat value that will be accepted for convergence}

\item{maxFailure}{The maximum number of convergence failures to accept}
}
\value{
Will either return a fitted stan object or the string "convergence
failed for X attempts" where X is the maxFailure parameter.
}
\description{
\code{safeFit} places a wrapper around rstan's code{stan} function allowing
for an R-hat convergence criteria to be specified. If after fitting this
criteria is not met by all parameters the model will be re-fitted a specified
number of time. If convergence is judged to have been achieved before this limit
is reached then the standard stan fit object, along with the number of attempts
made before convergence was achieved will be returned. If convergence is not
achieved in this number of steps then a string convergence failed for X
attempts" where X is the number of permitted failures.
}
\examples{
\dontrun{
safeFit(stanModel = "8schools.stan", stanData = schools_dat,
stanIter = 1000, stainChains = 4, maxRhat = 1.05)
}
}
