% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposedFunctions.R
\name{stanSim}
\alias{stanSim}
\title{Fits a stan model to multiple datasets and returns estimated values}
\usage{
stanSim(stanArgs = list(), simArgs = list(), returnArgs = list())
}
\arguments{
\item{stanArgs}{A list of model parameters to be taken by the \code{stan()} function.
If not specified then this functions defaults are used.}

\item{simArgs}{A list of parameters that control how the instances of \code{stan()}
are to be ran.}

\item{returnArgs}{A list of the parameters and summary statistics to return from all
fitted models.}
}
\value{
A dataframe of estimated values across all datasets
}
\description{
\code{stanSim} fits a specified stan model across multiple datasets and
collates and returns the estimated parameters for all models. The function
takes three sets of parameters (each provided in a seperate list); stan
modelling parameters, parameters governing the running of the multiple
stan instances, and a list of values to return (stan parameters along with
what summaries to include (e.g. percentiles)). The function allows for
stan instances to be ran in parallel, reducing run time on multicore
machines.
}
