% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansim_collection.R
\name{stansim_collection}
\alias{stansim_collection}
\title{Construct an S3 object of type stansim_collection}
\usage{
stansim_collection(collection_name, data, refitted, simulations)
}
\arguments{
\item{collection_name}{A name attached to the \code{stansim_collection}
object to help identify it. It is strongly recommended that an informative
name is assigned.}

\item{data}{The tidy dataframe containing data from all grouped simulations.}

\item{refitted}{A dataframe containing a row for every refitted dataset,
listing the simulation to which it belonged and the dataset name.}

\item{simulations}{A list of all simulation data other than estimated values
and the list of refitted data.}
}
\value{
An S3 object of class \code{stansim_collection}.
}
\description{
A constructor function for creating S3 objects of class
  \code{stansim_collection}. \code{stansim_collection} objects are the
  preferred means of managing and storing the results of multiple stan
  simulations ran using \code{stansim()}.
}
