% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansim_simulate.R
\name{stansim_simulate}
\alias{stansim_simulate}
\title{stansim_simulate placeholder}
\usage{
stansim_simulate(file, data_name = paste0("Simdata_", Sys.time()),
  save_dir = NULL, holding_data = NULL, sim_params = "all",
  param_values = NULL, datasets = 1, use_cores = 1, sim_drop = TRUE)
}
\description{
stansim_simulate placeholder
}
