% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansimObjects.R
\name{stansim_single}
\alias{stansim_single}
\title{Construct an S3 object of type stansim_single}
\usage{
stansim_single(sim_name, stansim_uni_list, start_time, end_time, stansim_seed)
}
\arguments{
\item{sim_name}{The name to be given to the simulation represented
by the \code{stansim_single} object.}

\item{stansim_uni_list}{A list of objects with S3 class
\code{stan_sim_uni}. This is an unexported class used to store the
outcomes of individual simulation runs internal to the \code{stansim}
function.}

\item{start_time}{System time when \code{stansim} was called.}

\item{end_time}{System time when the results from \code{stansim}
were returned.}

\item{stansim_seed}{The global seed for the \code{stansim} call.}
}
\value{
An S3 object of class \code{stansim_single} recording relevant
simulation data.
}
\description{
A constructor function for creating S3 Objects of
type \code{stansim_single}. \code{stansim_single} objects are the basic
unit of output from calls to the \code{stansim} function and collects
the specified data for all stan models fitted.
}
