% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.R
\name{stanSimChecker}
\alias{stanSimChecker}
\title{Validates input for stanSim function.}
\usage{
stanSimChecker(stanArgs, simArgs, returnArgs)
}
\arguments{
\item{stanArgs}{A list of model parameters to be taken by the \code{stan()} function.}

\item{simArgs}{A list of parameters that control how the instances of \code{stan()}
are to be ran.}

\item{returnArgs}{A list of the parameters and summary statistics to return from all
fitted models.}
}
\value{
Returns nothing if valid, stops the function and returns an error if any not.
}
\description{
\code{stanSimChecker} runs several tests on input to \code{stanSim()} to check for
validity early in the function.
}
