% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposedFunctions.R
\name{stanSimDefaults}
\alias{stanSimDefaults}
\title{Returns the default values for stanSim's three parameter lists}
\usage{
stanSimDefaults(getDefault = "All")
}
\arguments{
\item{getDefault}{Specifies which parameter list to access, defaults to all three
lists. Accepts values c("All", "stanArgsDefault", "simArgsDefault", "returnArgsDefault")}
}
\value{
A list of default parameter values
}
\description{
\code{stanSimDefaults} is used to view the default parameters that stanSim uses if not
overwritten by the user. All three lists are accessed by default or by entering
"All as the one parameter value. Individual default lists can be accessed by
specifying "stanArgsDefault", "simArgsDefault", or
"returnArgsDefault".
}
