% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write data from a \code{stansim_data} object to .rds files}
\usage{
write_data(object, path, data_name = NULL, recursive = FALSE)
}
\arguments{
\item{object}{An object of S3 class stansim_data.}

\item{path}{The name of the directory to write the .rds datafiles to. If the
directory does not exist it will be created. it's recommended that this is
the name of an empty or non-existant directory}

\item{data_name}{Controls the name stem for all saved .rds files. If NULL
this will be the \code{data_name} property from the provided
\code{stansim_data} object, otherwise it will be the provided character.
Default value is NULL.}

\item{recursive}{logical. Should elements of the path other than the last be
created? If true, like the Unix command mkdir -p.}
}
\description{
\code{write_data()} takes a \code{stansim_data} and a directory
  name and writes each of the objects' simulated datasets to an individual
  .rds file in the given directory.
}
\examples{
\dontrun{
}

}
