% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_refitted-methods.R
\name{extract_refitted}
\alias{extract_refitted}
\title{Extract identifiers of refitted items from rstansim objects}
\usage{
extract_refitted(object)
}
\arguments{
\item{object}{An object of S3 class stansim_simulation.}
}
\description{
Generic function for extracting refitted identifiers from
rstansim objects.
}
