% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_refitted-methods.R
\name{extract_refitted.stansim_simulation}
\alias{extract_refitted.stansim_simulation}
\title{Extract names of refitted datasets from a stansim_simulation object}
\usage{
\method{extract_refitted}{stansim_simulation}(object)
}
\arguments{
\item{object}{An object of S3 class stansim_simulation.}
}
\description{
Applied to an object of type stansim_simulation,
\code{extract_refitted} will return a vector of names of datasets
that have been refitted since the initial simulation.
}
