% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansimMethods.R
\name{print.stansim_simulation}
\alias{print.stansim_simulation}
\title{Print a summary for a stansim simulation object}
\usage{
\method{print}{stansim_simulation}(x, ...)
}
\arguments{
\item{x}{An object of S3 class stansim.}

\item{...}{other arguments not used by this method}
}
\description{
Print basic  information regarding a stansim simulation object,
including simulation title, time/date ran, number of models fitted, and parameters
and estimates recorded.
}
