% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitMethods.R
\name{refit}
\alias{refit}
\title{Refit specified datafiles in a stansim object}
\usage{
refit(object, ...)
}
\arguments{
\item{object}{An object of S3 class stansim_single.}

\item{...}{Arguments specifying datafiles to refit and control
of warning behaviour.}
}
\description{
Generic function for refitting specified datafiles in
rstansim objects. By default refits all datafiles, following a user
prompt given the potential computational cost (this can be turned off).
Otherwise only datafile names provided will be refit.
}
