% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename a stansim object}
\usage{
rename(object, new_name)
}
\arguments{
\item{object}{An object of S3 class stansim_simulation.}

\item{new_name}{New object name. Must be of type character.}
}
\description{
Generic function for refitting specified datafiles in
rstansim objects. By default refits all datafiles, following a user
prompt given the potential computational cost (this can be turned off).
Otherwise only datafile names provided will be refit.
}
