% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposedFunctions.R
\name{rename}
\alias{rename}
\title{Rename a stansim simulation object}
\usage{
rename(object, new_name)
}
\arguments{
\item{object}{An object of S3 class stansim_single.}

\item{new_name}{New object name. Must be of type character.}
}
\description{
Change the sim_name value of a \code{"stansim_single"}
object.
}
