% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposedFunctions.R
\name{stan_sim}
\alias{stan_sim}
\title{Fits a stan model to multiple datasets and returns estimated values}
\usage{
stan_sim(stan_args = list(), sim_data = NULL, calc_loo = FALSE,
  use_cores = 1L, parameters = ".*", estimates = c("2.5\%", "50\%",
  "97.5\%", "n_eff", "Rhat"), stan_warnings = "catch", cache = TRUE,
  stansim_seed = floor(runif(1, 1, 1e+05)), sim_name = paste0("Stansim_",
  start_time))
}
\arguments{
\item{stan_args}{A list of model parameters to be taken by
the \code{stan()} function. If not specified then the \code{stan()}
function defaults are used.}

\item{sim_data}{A list of strings pointing to the .csv files that
contain the simulation data. Only data that varies from model to
model must be specified here, data common to all models can be
specified as part of \code{stan_args}.}

\item{calc_loo}{If \code{TRUE} the model will attempt to extract fit
statistics using the loo package. If \code{TRUE} there must be
a valid log_lik quantity specified in the generated quantities
section of the stan model.}

\item{use_cores}{Number of cores to use when running in parallel.
Each stan model is fitted serially regardless of the number of chains
ran. However as the number of available cores increases more models
can be fit in parallel}

\item{parameters}{A set of characters indicating which parameters
should have estimates returned from the fitted models. Regular
expressions are used to extract parameters, so care has to be taken
with similarly named parameters (e.g. 'eta' and 'theta').}

\item{estimates}{The estimates that should be returned for each
model parameter. Currently only supports the standard 'summary'
parameters c('mean', 'se_mean',  'sd', '2.5%', '25%',  '50%',
'75%', '97.5%', 'n_eff', 'Rhat').}

\item{max_failures}{The maximum number of times that each model
will attempt refitting when its greatest R-hat statstics
is > \code{max_rhat}. If this number is exceeded a string will be returned
indicating convergence failure, rather than model estimates.}

\item{max_rhat}{A maximum cut-off for crudely assessing convergence
across fitted models. If any parameters in the fitted model have an
R-hat >  \code{max_rhat} then the model will either attempt re-fitting
or will return a string indicating lack of convergence (depending on
the number of fitting attempts made and the value of \code{max_failures}).
Setting this to a large value (e.g. 1000) will
efectively turn off the convergence assessment functionality.}
}
\value{
A dataframe of estimated values across all datasets.
}
\description{
\code{stan_sim} fits a specified stan model across multiple
datasets and collates and returns the estimated parameters for all models.
It fits the specified model to all datasets indicated and returns the
specified results for each model. It provides an optional set of controls
in case any fitted models do not appear to have converged, in which case
model fit will be re-attempted or a string indicating lack of fit will be
returned for that model. The function allows for stan instances to be
ran in parallel, reducing run the time taken on multicore machines.
}
