% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansim_simulate.R
\name{stansim_simulate}
\alias{stansim_simulate}
\title{Simulate datasets from a stan model}
\usage{
stansim_simulate(file, data_name = paste0("Simdata_", Sys.time()),
  input_data = NULL, sim_params = "all", param_values = NULL,
  datasets = 1, save_dir = NULL, return_object = TRUE, use_cores = 1,
  sim_drop = TRUE)
}
\arguments{
\item{file}{A character string or a connection that R supports specifying the Stan model specification in Stan's modeling language.}

\item{data_name}{A name attached to the \code{stansim_data} object to
help identify it. It is strongly recomended that an informative name is
assigned. If \code{save_dir} isn't NULL, this will also be the name stem for the saved .rds files.}

\item{input_data}{temp}

\item{sim_params}{temp}

\item{param_values}{temp}

\item{datasets}{The number of simulated datasets to produce.}

\item{save_dir}{If NULL}

\item{return_object}{if FALSE the no \code{stansim_data} object is returned. Use along with \code{save_dir} to write output data
that is too large to store and manipulate in memory.}

\item{use_cores}{Number of cores to use when running in parallel.}

\item{sim_drop}{temp}
}
\description{
\code{stansim_simulate()} takes a specified stan model and allows the user to simulate data from it
based on specified parameter values. The user then specified which data they wish to return and how many simulations
they wish to run. By default an object of class \code{stansim_data} will be returned, but if a character is provided to
\code{save_dir} then the data will be saved as individual .rds files in the specified directory.

To allow for simulated data to be directly fed into stan model that simulated them as input data, the sim_drop argument is provided.
If \code{sim_drop} is true then any stan data object with a name beginning with "sim_" wil have this string removed from it's name.
For example, the simulated data "sim_x" would be returned simply as "x". This helps avoid the issue of overlapping data names for both
input and output
}
